Create Traversal.cs
just the traversal script. The 9/24 NavMesh lecture recording covers the same thing and how to get set up with the AI navigation component. Basically you open the package manager and in the unity registry you install the AI Navigation package. Once it's installed, you select all of the items that would be obstacles (trees, rocks, ect.) and add the navmesh obstacle component to them. Then select the terrain object (the ground, plane, whatever you want to call it) and add the navmesh surface component to it. Then you want to bake the surface from that component interface. Finally you add the object you want to move and another object that will serve as a target. Attach the script to the first one and drag the target object to the target component of the script. Make sure that the object that you want to move is close to the ground, if it's too high up it won't move. Maybe tweak the height component (covered in the vid).